feat: move Searchbar to its own component
